api/requirements-old.txt
alpha-vantage==2.1.3
Brotli==1.0.7
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
dash==1.11.0
dash-core-components==1.9.1
dash-html-components==1.0.3
dash-renderer==1.4.0
dash-table==4.6.2
Flask==1.1.2
Flask-Compress==1.5.0
future==0.18.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.18.3
pandas==1.0.3
plotly==4.6.0
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2019.3
requests==2.23.0
retrying==1.3.3
six==1.14.0
urllib3==1.25.9
Werkzeug==1.0.1